refactor: seo type 위치 변경 및 header 컴포넌트 반환 jsx 수정
