+++ src/components/common/Header/index.tsx
@@ -6,8 +6,6 @@ import Gnb from '../Gnb';
 import { throttle } from '@/utils/throttle';
 
 function Header() {
-  if (typeof window === 'undefined') return;
-
   const [isHide, setIsHide] = useState(false);
   const [scroll, setScroll] = useState(window.scrollY);
