+++ src/types/PostItem.types.ts
@@ -22,13 +22,3 @@ export type PostPageItemProps = {
     frontmatter: PostFrontmatterType;
   };
 };
-
-export type SEOType = {
-  site: {
-    siteMetadata: {
-      title: string;
-      description: string;
-      author: string;
-    };
-  };
-};
