+++ src/types/SEO.types.ts
@@ -1,0 +1,9 @@
+export type SEOType = {
+  site: {
+    siteMetadata: {
+      title: string;
+      description: string;
+      author: string;
+    };
+  };
+};
